+++ src/routes/sale/escrow/EscrowUndepositTable.svelte
@@ -1,0 +1,225 @@
+<script lang="ts">
+  import { formatAddress } from "src/utils";
+  import { query } from "@urql/svelte";
+  import { formatUnits } from "ethers/lib/utils";
+  import { signerAddress } from "svelte-ethers-store";
+  import { getContext } from "svelte";
+  import IconLibrary from "../../../components/IconLibrary.svelte";
+  import { allUndepositQuery, myUndepositQuery } from "./escrow-queries";
+  import Switch from "src/components/Switch.svelte";
+  import EscrowUndepositModal from "./EscrowUndepositModal.svelte";
+  import { Contract } from "ethers";
+
+  const { open } = getContext("simple-modal");
+  export let salesContract, saleData, escrow: Contract;
+
+  let checked = true;
+
+  $: txQuery = checked ? allUndepositQuery : myUndepositQuery;
+
+  // init the queries
+  query(allUndepositQuery);
+  query(myUndepositQuery);
+
+  // setting the query variables
+  $myUndepositQuery.variables.saleAddress = salesContract.address.toLowerCase();
+  $myUndepositQuery.variables.depositor = $signerAddress.toLowerCase();
+
+  $allUndepositQuery.variables.saleAddress =
+    salesContract.address.toLowerCase();
+
+  // handling table refresh
+  const refresh = () => {
+    $txQuery.reexecute();
+  };
+</script>
+
+<div class="flex w-full flex-col gap-y-4">
+  <div class="flex flex-row justify-between">
+    <span class="text-lg font-semibold">Escrow Undeposit History</span>
+    <div class="flex flex-row items-center gap-x-4">
+      <span class="text-sm"
+        >Showing {#if !checked}only mine{:else}all transactions{/if}</span
+      >
+      <Switch bind:checked />
+      <span
+        class:animate-spin={$txQuery.fetching}
+        class="flex flex-col justify-center"
+        on:click={refresh}><IconLibrary icon="reload" /></span
+      >
+    </div>
+  </div>
+  {#if $txQuery.fetching}
+    Loading transactions...
+  {:else if $txQuery.error}
+    Something went wrong.
+  {:else if $txQuery?.data?.redeemableEscrowSupplyTokenDepositors.length}
+    <table class="table-auto w-full space-y-2 text-sm">
+      <tr class="border-b border-gray-600 uppercase text-sm">
+        <th class="text-gray-400 text-left pb-2 font-light">Depositor</th>
+        <th class="text-gray-400 text-left pb-2 font-light">Token Address</th>
+        <th class="text-gray-400 text-left pb-2 font-light">Symbol</th>
+        <th class="text-gray-400 text-left pb-2 font-light">Total Deposited</th>
+        <th class="text-gray-400 text-left pb-2 font-light">Remaining</th>
+      </tr>
+      {#each $txQuery.data.redeemableEscrowSupplyTokenDepositors as data}
+        <tr>
+          <td> {formatAddress(data.depositorAddress)} </td>
+          <td>
+            {formatAddress(data.token.id)}
+          </td>
+          <td>
+            {data.token.symbol}
+          </td>
+          <td class="py-2">
+            {Number(
+              (+formatUnits(data.totalDeposited, data.token.decimals)).toFixed(
+                4
+              )
+            )}
+            {data.token.symbol}
+          </td>
+          <td class="py-2">
+            {Number(
+              (+formatUnits(data.totalRemaining, data.token.decimals)).toFixed(
+                4
+              )
+            )}
+            {data.token.symbol}
+          </td>
+          <td class="py-2 text-right">
+            <span
+              class="underline cursor-pointer text-gray-400 mr-4"
+              on:click={() => {
+                open(EscrowUndepositModal, {
+                  escrow,
+                  data,
+                  salesContract,
+                  saleData,
+                });
+              }}
+            >
+              Undeposit
+            </span>
+          </td>
+        </tr>
+
+        <!-- {#if data.deposits.length > 1}
+          {#each data.deposits as undeposit}
+            <tr>
+              <td>
+                {formatAddress(undeposit.depositorAddress)}
+              </td>
+              <td>
+                {formatAddress(data.token.id)}
+              </td>
+              <td>
+                {data.token.symbol}
+              </td>
+              <td class="py-2">
+                {Number(
+                  (+formatUnits(
+                    undeposit.tokenAmount,
+                    data.token.decimals
+                  )).toFixed(4)
+                )}
+                {data.token.symbol}
+              </td>
+              <td class="py-2">
+                {Number(
+                  (+formatUnits(
+                    data.totalDeposited,
+                    data.token.decimals
+                  )).toFixed(4)
+                )}
+                {data.token.symbol}
+              </td>
+              <td class="py-2">
+                {Number(
+                  (+formatUnits(
+                    data.totalRemaining,
+                    data.token.decimals
+                  )).toFixed(4)
+                )}
+                {data.token.symbol}
+              </td>
+              <td class="py-2 text-right">
+                {#if undeposit.depositorAddress == $signerAddress.toLowerCase()}
+                  <span
+                    class="underline cursor-pointer text-gray-400 mr-4"
+                    on:click={() => {
+                      open(EscrowUndepositModal, {
+                        escrow,
+                        data,
+                        salesContract
+                      });
+                    }}
+                  >
+                    Undeposit
+                  </span>
+                {/if}
+              </td>
+            </tr>
+          {/each}
+        {:else}
+          <tr>
+            <td>
+              {formatAddress(data.deposits[0].depositorAddress)}
+            </td>
+            <td>
+              {formatAddress(data.token.id)}
+            </td>
+            <td>
+              {data.token.symbol}
+            </td>
+            <td class="py-2">
+              {Number(
+                (+formatUnits(
+                  data.deposits[0].tokenAmount,
+                  data.token.decimals
+                )).toFixed(4)
+              )}
+              {data.token.symbol}
+            </td>
+            <td class="py-2">
+              {Number(
+                (+formatUnits(
+                  data.totalDeposited,
+                  data.token.decimals
+                )).toFixed(4)
+              )}
+              {data.token.symbol}
+            </td>
+            <td class="py-2">
+              {Number(
+                (+formatUnits(
+                  data.totalRemaining,
+                  data.token.decimals
+                )).toFixed(4)
+              )}
+              {data.token.symbol}
+            </td>
+            <td class="py-2 text-right">
+              {#if data.deposits[0].depositorAddress == $signerAddress.toLowerCase()}
+                <span
+                  class="underline cursor-pointer text-gray-400 mr-4"
+                  on:click={() => {
+                    open(EscrowUndepositModal, {
+                      escrow,
+                      data,
+                      salesContract,
+                    });
+                  }}
+                >
+                  Undeposit
+                </span>
+              {/if}
+            </td>
+          </tr>
+        {/if} -->
+      {/each}
+    </table>
+  {:else}
+    You haven't made any transactions.
+  {/if}
+</div>
