+++ src/routes/sale/escrow/RedeemableDeposit.svelte
@@ -1,7 +1,7 @@
 <script lang="ts">
   import { BigNumber, Contract, ethers, Signer } from "ethers";
-  import Button from "../../components/Button.svelte";
-  import FormPanel from "../../components/FormPanel.svelte";
+  import Button from "../../../components/Button.svelte";
+  import FormPanel from "../../../components/FormPanel.svelte";
   import { getContext } from "svelte";
   import DepositModal from "./DepositModal.svelte";
 
@@ -10,14 +10,12 @@
   export let saleData;
   export let sale: Contract;
   export let escrow: Contract;
-  
 </script>
 
 <FormPanel heading="Deposit Escrow">
-      <Button
-        on:click={() => {
-          open(DepositModal, { sale, saleData, escrow });
-        }}>Deposit</Button
-      >
-
+  <Button
+    on:click={() => {
+      open(DepositModal, { sale, saleData, escrow });
+    }}>Deposit</Button
+  >
 </FormPanel>
